# Adding scaling throttling to avoid rebalancing.