Dynamic scale does not work well with Event Hubs because of partition owner rebalancing. The Scale Controller votes to scale every 10 seconds. A scaling operation will assign/unassign workers, which leads to rebalancing. Rebalancing causes performance drops and duplicate executions.

To avoid frequent rebalancing, we want to introduce new throttling logic.

[Old throttling logic]
Throttle scale down if there was a scale up within last 60 seconds.

[New throttling logic]
Throttle scale down if there was a scale operation or the same target worker count was voted within 180 seconds.
Throttle scale down if there was a scale up in the last 60 seconds.
Throttle scale up if backlog is less than 30% of the execution rate.
On scale down, pick the maximum of the following: the highest vote from the history, or one step back from the sorted valid worker count.
Allow scale down to 0 if there were no executions in the last 60 seconds.