@jsquire 2024-01-02
/azp run net - eventhub - functions - tests