@jsquire 2024-01-03
/azp run net - eventhub - functions - tests